+++ src/utilities/include/par.h
@@ -51,7 +51,7 @@ enum class ExecutionPolicy {
 // - Sequential for small workload,
 // - Parallel (CPU) for medium workload,
 // - GPU for large workload if available.
-inline constexpr ExecutionPolicy autoPolicy(int size) {
+inline constexpr ExecutionPolicy autoPolicy(size_t size) {
   // some random numbers
   if (size <= (1 << 12)) {
     return ExecutionPolicy::Seq;
